LayoutTests/http/tests/app-privacy-report/new-window-isnotappinitiated.html

<!DOCTYPE html><!-- webkit-test-runner [ isAppInitiated=false ] -->
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Make sure that windows opened via window.open are attributed correctly");

window.jsTestIsAsync = true;

onload = function() {
    window.name = "opener";
    w = open("resources/new-window-isnotappinitiated-win.html", "");
}    
</script>
</body>
</html>
